Auto-detect when authentication is needed
When a request is made and we don't have an API token configured,
trying sending an unauthenticated API request.  If we get back a 401
Unauthorized, Automatic Login is not enabled and we ``authenticate``
to get an API token from a username/password combo.  Once we get the
API token to use, add it to the original request and send it on its
way.  If we don't get a 401, Automatic Login is enabled and we don't
need an API token and can send the original request unaltered.