Lock marshmallow package requirement to <3.0.0
Modify our requirements.txt file to prevent usage of version 3.0.0 or
newer of the marshallmallow package which is used by cdrouter.py for
JSON serialization/deserialization. The 3.x release includes many
breaking changes, in addition to dropping support for Python 2 and
Python < 3.5.  See the upgrade guide here:

https://marshmallow.readthedocs.io/en/stable/upgrading.html

As we have many users that are still using Python 2, this is most
likely the best course of action for the time being.