Add docker.sh for managing releases using Docker
Add a new docker.sh script which makes it possible manage releases and
run cdrouter.py inside a Docker container.  This makes it much simpler
to test out cdrouter.py in a clean environment using different
versions of Python.  Note that this commit does not affect users of
cdrouter.py in any way and is only meant to help maintainers of this
package.