Fix TagsService.edit method when renaming a tag
When renaming a tag, a new `old_name` parameter must be passed to the
`TagsService.edit` method.  This parameter defaults to `None` and is
optional unless the caller is renaming a tag.  The method needs to
know the old name of the tag so that is can construct the correct URL
for the request.  See
https://support.qacafe.com/cdrouter-web-api/tags#edit-a-tag for more
details.