Bump timeout for non-auth requests
Bump up the timeout for non-authentication API requests to 10min.
Some endpoints can take several minutes to return a response (i.e. the
upgrade endpoint).  The main purpose of the timeout is to ensure a
script won't hang indefinitely, therefore we use a conservative
timeout of 10min which should be long enough for even slow systems.