Upgrade to Marshmallow 3.18.0, drop support for Python 2.x
Upgrade cdrouter.py to use Marshmallow v3.3 and up, following changes
as described in:

https://marshmallow.readthedocs.io/en/stable/upgrading.html#upgrading-to-3-0
https://marshmallow.readthedocs.io/en/stable/upgrading.html#upgrading-to-3-3

Note that marshmallow v3.x only supports Python >= 3.5, therefore
cdrouter.py no longer supports Python 2 or Python < 3.5.

We no longer need our custom `DateTime` (from
`cdrouter/cdr_datetime.py`) or `DictField` (from
`cdrouter/cdr_dictfield.py`) classes as we can now use Marshmallow's
built-in `DateTime` and `Dict` field types:

https://marshmallow.readthedocs.io/en/stable/marshmallow.fields.html#marshmallow.fields.DateTime
https://marshmallow.readthedocs.io/en/stable/marshmallow.fields.html#marshmallow.fields.Dict

A `**kwargs` argument was added to many `post_load` methods in schema
classes, this was done following the following upgrade guide entry:

https://marshmallow.readthedocs.io/en/stable/upgrading.html#decorated-methods-and-handle-error-receive-many-and-partial

without which Marshmallow throws "unexpected more keyword" errors.
Because this argument is needed but not used, the pylint
`unused-argument` warning is disabled for each of these `post_load`
methods.

Marshmallow now throws an error when deserializing JSON if it finds a
field that is not represented in the schema.  This lead to the
addition of a number of missing fields in certain schemas:

- `type` was added to `captures.Capture`

- `keylog` was added to `testresults.TestResult`

- `more` was added to `testresults.Log`.

- `id`, `license_type`, `killtime`, `nag`, os_type` and `license_info`
  were added to `testsuites.Info`.

All pylint warnings have been resolved or disabled and all unit tests
are passing.