Build cdrouter.py Docker image and push to registry in CI job
Add new `docker` CI job to our CI pipeline which only runs if the
commit was tagged with a tag matching the `vX.Y.Z` format used for
official cdrouter.py releases.  When run, the CI job builds a Python
3.8-based Docker image containing cdrouter.py pre-installed and pushes
it to the GitLab Docker container registry for the project.  The
following image paths can be used with `docker run` to run containers
using the images:

    registry.gitlab.com/qacafe/cdrouter/cdrouter.py/cdrouter.py:X.Y.Z
    registry.gitlab.com/qacafe/cdrouter/cdrouter.py/cdrouter.py:latest

If run with no arguments to `docker run`, the images drop the user
into the interactive shell provided by running `python3` with no
arguments.  To run a cdrouter.py script using the images, mount the
directory containing the script on the host to a volume such as `/src`
in the container and pass the path as the first argument to `docker
run`.  For example, to run the cdrouter.py script `script.py` which is
in the CWD, run:

    docker run --rm -it -v $(pwd):/src \
        registry.gitlab.com/qacafe/cdrouter/cdrouter.py/cdrouter.py:latest /src/script.py