[sc-20602] Fix flaky JobService unit tests
Fix flaky unit tests in `tests/test_jobs.py` which test the
`JobService` class, the methods under `c.jobs`.  These tests were
often failing because the jobs launched into the job queue began
running or even finishing too quickly whereas the tests expected the
jobs to remain in the `pending` state.  This has been fixed by
launching all jobs with a `run_at` field set to a date in the year
9999.  This ensures that CDRouter's scheduler will not launch the jobs
and therefore ensures that the jobs will remain in the jobs queue in
the `pending` state as expected by the tests.