Mark flaky tests to be rerun
Using the `pytest-rerunfailures`
plugin (https://github.com/pytest-dev/pytest-rerunfailures), mark the
flaky unit test `test_configs.py::TestConfigs::test_get_networks` with
`@pytest.mark.flaky(reruns=5)` to indicate that it should be rerun up
to 5 times in the case of intermittent failures.  If the machine
running the cdrouter.py unit tests is heavily loaded, this test can
sometimes fail because the `cdrouter-cli` command the API endpoint
relies on sometimes times out when trying to validate the value of the
`docsisConfig` testvar, which is given 10s to return.