[sc-25225] Ignore unknown fields in API responses
In sc-6359, we upgraded from Marshmallow 2.x to 3.x, which brought a
number of changes.  One change which was unfortunately not caught was
that fields not defined in cdrouter.py's schemas caused Marshmallow to
raise a `ValidationError`.  Thus, starting in cdrouter.py v0.7.0 (the
first version based on Marshmallow 3.x), if an API response included a
field that cdrouter.py didn't know about, a `ValidationError` would
occur.  This would oftentimes happen if the CDRouter system was
upgraded to a newer version without also updating cdrouter.py.  We
don't want to force people to upgrade cdrouter.py, therefore the
current behavior of Marshmallow is not ideal.  Fortunately,
Marshmallow provides a way to revert to the 2.x behavior:

https://marshmallow.readthedocs.io/en/stable/upgrading.html#schemas-raise-validationerror-when-deserializing-data-with-unknown-keys

This has been done by adding an `unknown=EXCLUDE` argument to all
`schema.load()` calls, which should cause Marshmallow to ignore
unknown fields as it did in the 2.x series.  A number of unit tests
have been added to ensure this.