Fetch CDRouter CI license from QA Cafe Admin Lounge
Instead of using the static license stored in the `CDR_DOCKER_LICENSE`
GitLab CI variable, fetch a CDRouter CI license from the QA Cafe Admin
Lounge to use at the start of every `test` CI job.  This license is a
subscription CDRouter license with 5 execution instances and all
add-ons enabled and is always valid for the next 24 hours.