Stop & remove CDRouter container if error on startup
Modify the `cdrouter` pytest fixture used by cdrouter.py unit tests to
bring up a fresh CDRouter container to stop & remove the CDRouter
container if an error occurs while bringing it up.  This could be due
to the fixture being unable to determine the address/port of the
container, the `/usr/cdrouter/etc/cdrouter.sh` failing to bring up the
CDRouter service within the container, or the call to the system time
API endpoint `c.system.time()` failing to report that we are able to
make a successful API call to the container.